feat: implement answered elsewhere logic in history processing
