+++ root/usr/lib/node/nethcti-server/plugins/com_history_rest/plugins_rest/historycall.js
@@ -4,9 +4,8 @@
  * @module com_history_rest
  * @submodule plugins_rest
  */
-var fs = require('fs');
-var net = require('net');
 var path = require('path');
+var answeredElsewhereLive = require('../answered_elsewhere_live');
 
 /**
  * The module identifier used by the logger.
@@ -83,196 +82,6 @@ var compStaticHttp;
  * @private
  */
 var compAstProxy;
-var amiConfigCache;
-
-function loadAmiConfig() {
-  if (amiConfigCache) {
-    return amiConfigCache;
-  }
-
-  amiConfigCache = JSON.parse(fs.readFileSync('/etc/nethcti/asterisk.json', 'utf8'));
-  return amiConfigCache;
-}
-
-function parseAmiMessage(message) {
-  var parsed = {};
-
-  message.split(/\r\n/).forEach(function (line) {
-    var separatorIndex = line.indexOf(':');
-    if (separatorIndex === -1) {
-      return;
-    }
-
-    var key = line.substring(0, separatorIndex).trim().toLowerCase();
-    var value = line.substring(separatorIndex + 1).trim();
-    parsed[key] = value;
-  });
-
-  return parsed;
-}
-
-function getActiveLinkedids(cb) {
-  try {
-    var amiConfig = loadAmiConfig();
-    var socket = net.createConnection({
-      host: amiConfig.host || 'localhost',
-      port: parseInt(amiConfig.port, 10) || 5038
-    });
-    var actionId = 'historycall-' + Date.now();
-    var buffer = '';
-    var loggedIn = false;
-    var completed = false;
-    var activeLinkedids = {};
-    var timeout = setTimeout(function () {
-      cleanup(new Error('AMI CoreShowChannels timeout'));
-    }, 5000);
-
-    function finish(err) {
-      if (completed) {
-        return;
-      }
-
-      completed = true;
-      clearTimeout(timeout);
-
-      if (socket && !socket.destroyed) {
-        socket.end();
-      }
-
-      cb(err, activeLinkedids);
-    }
-
-    function cleanup(err) {
-      if (socket && !socket.destroyed) {
-        socket.destroy();
-      }
-
-      finish(err);
-    }
-
-    socket.on('connect', function () {
-      socket.write(
-        'Action: Login\r\n' +
-        'Username: ' + amiConfig.user + '\r\n' +
-        'Secret: ' + amiConfig.pass + '\r\n' +
-        'Events: off\r\n\r\n'
-      );
-    });
-
-    socket.on('data', function (chunk) {
-      buffer += chunk.toString();
-
-      while (buffer.indexOf('\r\n\r\n') !== -1) {
-        var separatorIndex = buffer.indexOf('\r\n\r\n');
-        var rawMessage = buffer.substring(0, separatorIndex);
-        buffer = buffer.substring(separatorIndex + 4);
-
-        if (!rawMessage.trim()) {
-          continue;
-        }
-
-        var message = parseAmiMessage(rawMessage);
-
-        if (!loggedIn) {
-          if (message.response === 'Success') {
-            loggedIn = true;
-            socket.write(
-              'Action: CoreShowChannels\r\n' +
-              'ActionID: ' + actionId + '\r\n\r\n'
-            );
-          } else if (message.response === 'Error') {
-            cleanup(new Error('AMI login failed: ' + (message.message || 'unknown error')));
-          }
-          continue;
-        }
-
-        if (message.actionid && message.actionid !== actionId) {
-          continue;
-        }
-
-        if (message.event === 'CoreShowChannel' && message.linkedid) {
-          activeLinkedids[message.linkedid] = true;
-        } else if (message.event === 'CoreShowChannelsComplete') {
-          socket.write('Action: Logoff\r\n\r\n');
-          finish(null);
-        } else if (message.response === 'Error') {
-          cleanup(new Error('AMI CoreShowChannels failed: ' + (message.message || 'unknown error')));
-        }
-      }
-    });
-
-    socket.on('error', function (err) {
-      cleanup(err);
-    });
-
-    socket.on('end', function () {
-      finish(null);
-    });
-
-  } catch (err) {
-    cb(err);
-  }
-}
-
-function promoteAnsweredElsewhereRows(results, cb) {
-  try {
-    if (!results || !Array.isArray(results.rows) || results.rows.length === 0) {
-      cb(null, results);
-      return;
-    }
-
-    var plainRows = results.rows.map(function (row) {
-      if (row && typeof row.get === 'function') {
-        return row.get({ plain: true });
-      }
-      if (row && row.dataValues) {
-        return row.dataValues;
-      }
-      return row;
-    });
-
-    var candidateRows = plainRows.filter(function (row) {
-      return row &&
-        row.linkedid &&
-        (
-          row.queue ||
-          (typeof row.channel === 'string' && row.channel.indexOf('@from-queue-') !== -1) ||
-          row.lastapp === 'Queue'
-        ) &&
-        ['NO ANSWER', 'BUSY', 'FAILED'].indexOf(row.disposition) !== -1;
-    });
-
-    if (candidateRows.length === 0) {
-      cb(null, results);
-      return;
-    }
-
-    getActiveLinkedids(function (err, activeLinkedids) {
-      if (err) {
-        logger.log.warn(IDLOG, 'failed to fetch active linkedids from AMI: ' + err.message);
-        cb(null, results);
-        return;
-      }
-      plainRows.forEach(function (row) {
-        if (activeLinkedids[row.linkedid]) {
-          if (['NO ANSWER', 'BUSY', 'FAILED'].indexOf(row.disposition) !== -1 &&
-            (
-              row.queue ||
-              (typeof row.channel === 'string' && row.channel.indexOf('@from-queue-') !== -1) ||
-              row.lastapp === 'Queue'
-            )) {
-            row.disposition = 'ANSWERED_ELSEWHERE';
-          }
-        }
-      });
-
-      results.rows = plainRows;
-      cb(null, results);
-    });
-  } catch (err) {
-    cb(err);
-  }
-}
 
 /**
  * Sets the asterisk proxy architect component.
@@ -951,7 +760,7 @@ function setCompAuthorization(ca) {
               if (err1) {
                 throw err1;
               } else {
-                promoteAnsweredElsewhereRows(results, function (err2, promotedResults) {
+                answeredElsewhereLive.promoteAnsweredElsewhereRows(results, logger, IDLOG, function (err2, promotedResults) {
                   try {
                     if (err2) {
                       throw err2;
